requirements.txt.txt
Django==1.11.18
djangorestframework==3.9.2
django-filter==2.1.0
django-rest-auth==0.9.3
pytz==2018.9
markdown==3.0.1
psycopg2==2.7.7
pillow==5.4.1
chardet==3.0.4 
idna==2.8 
requests==2.21.0 
urllib3==1.24.1 
certifi==2019.3.9
django-widget-tweaks==1.4.3
python-memcached==1.59 
six==1.12.0
asgi-redis==1.4.3 
asgiref==1.1.2 
msgpack-python==0.5.6 
redis==2.10.6
websocket-client==0.55.0
pathlib==1.0.1
incremental==17.5.0 
Automat==0.7.0
PyHamcrest==1.9.0
attrs==19.1.0
###async-timeout==3.0.1 
autobahn==19.3.1 
channels==0.17.3 
daphne==2.2.5
twisted==18.9.0 
txaio==18.8.1
constantly==15.1.0
hyperlink==18.0.0
zope.interface==4.6.0
gunicorn==19.9.0
django-channels-presence==0.0.9
Celery==4.3.0
amqp==2.4.2
billiard==3.6.0.0
kombu==4.5.0
vine==1.3.0
channels-redis==2.4.0
asgiref==1.1.2
channels==1.1.7
daphne==1.4.2
PyJWT==1.7.1
djangorestframework-jwt==1.11.0


